Cleanup `PlaylistWebLoader` `WKNavigationDelegate`
